tests-rls: add new benchmark entry for test_term_payloads_eval
