+++ tlspuffin/src/test_utils.rs
@@ -1,11 +1,30 @@
+use std::cmp::max;
+use std::collections::HashSet;
 use std::time::Duration;
 
+use itertools::Itertools;
+use puffin::agent::AgentName;
+use puffin::algebra::dynamic_function::DescribableFunction;
+use puffin::algebra::signature::FunctionDefinition;
+use puffin::algebra::{DYTerm, Term, TermType};
+use puffin::error::Error;
 use puffin::execution::{ExecutionStatus, ForkedRunner, Runner, TraceRunner};
+use puffin::fuzzer::term_zoo::TermZoo;
+use puffin::fuzzer::utils::{choose, find_term_by_term_path_mut, Choosable, TermConstraints};
+use puffin::libafl_bolts::bolts_prelude::{Rand, RomuDuoJrRand};
+use puffin::protocol::{ProtocolBehavior, ProtocolTypes};
 use puffin::put::PutDescriptor;
-use puffin::trace::{Spawner, Trace};
+use puffin::trace::Action::Input;
+use puffin::trace::{InputAction, Spawner, Step, Trace, TraceContext};
 
 use crate::protocol::{TLSProtocolBehavior, TLSProtocolTypes};
 use crate::put_registry::tls_registry;
+use crate::tls::fn_impl::{
+    fn_certificate_transcript, fn_client_finished_transcript, fn_decrypt_application,
+    fn_decrypt_multiple_handshake_messages, fn_server_finished_transcript,
+    fn_server_hello_transcript,
+};
+use crate::tls::TLS_SIGNATURE;
 
 pub fn default_runner_for(put: impl Into<PutDescriptor>) -> Runner<TLSProtocolBehavior> {
     let registry = tls_registry();
@@ -249,4 +268,307 @@ pub mod prelude {
     pub use crate::put_registry::{for_puts, tls_registry};
     pub use crate::test_utils::tcp::*;
     pub use crate::test_utils::{default_runner_for, expect_trace_crash};
+}
+
+/// Functions that are known to fail to be adversarially generated
+pub fn ignore_gen() -> HashSet<String> {
+    [
+        // As expected, attacker cannot use them as there is no adversarial
+        // '*Transcript*', which are required as argument
+        fn_server_finished_transcript.name(),
+        fn_client_finished_transcript.name(),
+        fn_server_hello_transcript.name(),
+        fn_certificate_transcript.name(),
+    ]
+    .iter()
+    .map(|fn_name| fn_name.to_string())
+    .collect::<HashSet<String>>()
+}
+
+/// Functions that are known to fail to be evaluated (without payloads)
+pub fn ignore_eval() -> HashSet<String> {
+    let mut ignore_gen = ignore_gen();
+    let ignore_eval = [
+        // Those 2 are the function symbols for which we can generate a term but all fail to
+        // DY_execute! Indeed, the HandshakeHash that is fed as argument must be
+        // computed in a very specific way! We might give known,valid hash-transcript to help?
+        fn_decrypt_application.name(),
+        fn_decrypt_multiple_handshake_messages.name(),
+    ]
+    .iter()
+    .map(|fn_name| fn_name.to_string())
+    .collect::<HashSet<String>>();
+    ignore_gen.extend(ignore_eval);
+    ignore_gen
+}
+
+/// Functions that are known to fail to be adversarially generated
+pub fn ignore_add_payload() -> HashSet<String> {
+    let mut ignore_eval = ignore_eval();
+    let ignore_pay: HashSet<String> = [
+        // No additional failures
+        // fn_find_server_certificate_verify.name(),
+        // fn_find_server_certificate_verify.name(),
+        // fn_find_server_finished.name(),
+    ]
+    .iter()
+    .map(|fn_name: &&str| fn_name.to_string())
+    .collect::<HashSet<String>>();
+    ignore_eval.extend(ignore_pay);
+    ignore_eval
+}
+/// Parametric test for testing operations on terms (closure `test_map`, e.g., evaluation) through
+/// the generation of a zoo of terms
+pub fn zoo_test<Ft>(
+    mut test_map: Ft,
+    mut rand: RomuDuoJrRand,
+    how_many: usize, // number of terms to generate for each function symbol (at root position)
+    stop_on_success: bool, /* do not test further term if its function at root position was
+                      * already positively tested */
+    stop_on_error: bool, /* for each function, stop testing further terms if an error is
+                          * encountered */
+    filter: Option<&FunctionDefinition<TLSProtocolTypes>>,
+    ignored_functions: &HashSet<String>,
+) -> bool
+where
+    Ft: FnMut(
+        &Term<TLSProtocolTypes>,
+        &TraceContext<TLSProtocolBehavior>,
+        &mut RomuDuoJrRand,
+    ) -> Result<(), Error>,
+{
+    let tls_registry = tls_registry();
+    let spawner = Spawner::new(tls_registry.clone());
+    let ctx = TraceContext::new(spawner);
+
+    let all_functions_shape = TLS_SIGNATURE.functions.to_owned();
+    let number_functions = all_functions_shape.len();
+    let mut number_terms = 0;
+    let mut number_success = 0;
+    let mut number_failure = 0;
+    let mut number_failure_on_ignored = 0;
+    let mut successful_functions = vec![];
+
+    let bucket_size = 200;
+    for f in &all_functions_shape {
+        if filter.is_none() || (filter.is_some() && filter.unwrap().0.name == f.0.name) {
+            'outer: for i in 0..(how_many / bucket_size) {
+                let bucket_size_step = if i < how_many / bucket_size - 1 {
+                    bucket_size
+                } else {
+                    how_many - bucket_size * (how_many / bucket_size - 1)
+                };
+                let zoo_f = TermZoo::<TLSProtocolTypes>::generate_many(
+                    &TLS_SIGNATURE,
+                    &mut rand,
+                    bucket_size_step,
+                    Some(&f),
+                );
+                let terms_f = zoo_f.terms();
+                if terms_f.len() != how_many {
+                    log::error!(
+                        "Failed to generate {bucket_size_step} terms (only {}) for function {}.",
+                        terms_f.len(),
+                        f.0.name
+                    );
+                }
+                number_terms += terms_f.len();
+
+                for term in terms_f.iter() {
+                    match test_map(term, &ctx, &mut rand) {
+                        Ok(_) => {
+                            successful_functions.push(term.name().to_string());
+                            number_success += 1;
+                            if stop_on_success {
+                                break 'outer;
+                            }
+                        }
+                        Err(e) => {
+                            if ignored_functions.contains(term.name()) {
+                                log::debug!("[Ignored function] Failed to test_map term {term} with error {e}. ");
+                                number_failure_on_ignored += 1;
+                            } else {
+                                log::error!("[Not ignored function] Failed to test_map term {term} with error {e}. ");
+                                number_failure += 1;
+                                if stop_on_error {
+                                    break 'outer;
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+    let all_functions = all_functions_shape
+        .iter()
+        .map(|(shape, _)| shape.name.to_string())
+        .collect::<HashSet<String>>();
+
+    let mut successful_functions = successful_functions
+        .into_iter()
+        .collect::<HashSet<String>>();
+    let successful_functions_tested = successful_functions.clone();
+    successful_functions.extend(ignored_functions.clone());
+
+    let difference = all_functions.difference(&successful_functions);
+    let difference_inverse = successful_functions_tested.intersection(&ignored_functions);
+
+    log::debug!("[zoo_test] ignored_functions: {:?}\n", &ignored_functions);
+    log::error!("[zoo_test] Diff: {:?}", &difference);
+    log::error!(
+        "[zoo_test] Intersection with ignored: {:?}",
+        &difference_inverse
+    );
+    log::error!(
+        "[zoo_test] Stats: how_many: {how_many}, stop_on_success: {stop_on_success}, stop_on_error: {stop_on_error}\n\
+        --> number_functions: {}, number_terms: {}, number_success: {}, number_failure: {}, number_failure_on_ignored: {}\n\
+        --> Successfully built (out of {:?} functions): {:?}",
+        number_functions,
+        number_terms,
+        number_success,
+        number_failure,
+        number_failure_on_ignored,
+        &all_functions.len(),
+        &successful_functions_tested.len()
+    );
+    (difference.count() == 0) && (difference_inverse.count() == 0)
+}
+
+pub fn add_payloads_randomly<
+    PT: ProtocolTypes,
+    R: Rand,
+    PB: ProtocolBehavior<ProtocolTypes = PT>,
+>(
+    t: &mut Term<PT>,
+    rand: &mut R,
+    ctx: &TraceContext<PB>,
+) {
+    let all_subterms: Vec<&Term<PT>> = t.into_iter().collect_vec();
+    let nb_subterms = all_subterms.len() as i32;
+    let mut i = 0;
+    let nb = (1..max(4, nb_subterms / 3))
+        .collect::<Vec<i32>>()
+        .choose(rand)
+        .unwrap()
+        .to_owned();
+    log::debug!(
+        "Adding {nb} payloads for #subterms={nb_subterms}, max={} in term: {t}...",
+        max(2, nb_subterms / 5)
+    );
+    let mut tries = 0;
+    // let nb = 1;
+    while i < nb {
+        tries += 1;
+        if tries > nb * 100 {
+            log::error!("Failed to add the payloads after {} attempts", tries);
+            break;
+        }
+        if let Ok(()) = add_one_payload_randomly(t, rand, ctx) {
+            i += 1;
+        }
+    }
+}
+
+/// Sanity check for the next test
+pub fn test_pay<PT: ProtocolTypes>(term: &Term<PT>) {
+    rec_inside(term, false, term);
+    pub fn rec_inside<PT: ProtocolTypes>(
+        term: &Term<PT>,
+        already_found: bool,
+        whole_term: &Term<PT>,
+    ) {
+        let already_found = already_found || !term.is_symbolic();
+        match &term.term {
+            DYTerm::Variable(_) => {}
+            DYTerm::Application(_, sub) => {
+                for ti in sub {
+                    if already_found && !ti.is_symbolic() {
+                        panic!("Eheh, found one! Sub: {ti},\n whole_term: {whole_term}")
+                    } else {
+                        rec_inside(ti, already_found, whole_term)
+                    }
+                }
+            }
+        }
+    }
+}
+
+pub fn add_one_payload_randomly<
+    PT: ProtocolTypes,
+    R: Rand,
+    PB: ProtocolBehavior<ProtocolTypes = PT>,
+>(
+    t: &mut Term<PT>,
+    rand: &mut R,
+    ctx: &TraceContext<PB>,
+) -> Result<(), Error> {
+    let trace = Trace {
+        descriptors: vec![],
+        steps: vec![Step {
+            agent: AgentName::new(),
+            action: Input(InputAction {
+                precomputations: vec![],
+                recipe: t.clone(),
+            }),
+        }],
+        prior_traces: vec![],
+    };
+    if let Some((st_, (step, mut path))) = choose(
+        &trace,
+        TermConstraints {
+            // as for Make_message.mutate
+            no_payload_in_subterm: false,
+            not_inside_list: false, // should be true, TODO: fix this
+            weighted_depth: false,  // should be true, TODO: fix this
+            ..TermConstraints::default()
+        },
+        rand,
+        // Tests by varying TermConstraints (diff=2 corresponds to fn_derive_psk.name(),
+        // fn_get_ticket.name()) Before   not_inside_list: true,
+        // no_payload_in_subterm: true,     weighted_depth: true, : number_shapes:
+        // 201, number_terms: 78800, eval_count: 182, count_payload_fail: 53, count_lazy_fail:
+        // 3668, count_any_encode_fail: 0 DIF=5
+
+        // Default (all false): number_shapes: 201, number_terms: 78800, eval_count: 185,
+        // count_payload_fail: 22, count_lazy_fail: 3839, count_any_encode_fail: 0
+        // Dif=2
+        //
+        // MAke_message: true, false, true
+        //  number_shapes: 201, number_terms: 78800, eval_count: 181, count_payload_fail: 34,
+        // count_lazy_fail: 3957, count_any_encode_fail: 0 Diff = 6
+
+        // MAke_message + inside: false, false, true
+        // number_shapes: 201, number_terms: 78800, eval_count: 183, count_payload_fail: 23,
+        // count_lazy_fail: 3260, count_any_encode_fail: 0 Diff = 4
+
+        // weighted_Depth = false
+        // number_shapes: 201, number_terms: 78800, eval_count: 184, count_payload_fail: 31,
+        // count_lazy_fail: 4018, count_any_encode_fail: 0 Diff = 3
+    ) {
+        let st = find_term_by_term_path_mut(t, &mut path).unwrap();
+        if let Ok(()) = st.make_payload(ctx) {
+            log::debug!("Added payload for subterm at path {path:?}, step{step},\n - sub_term: {st_}\n  - whole_term {trace}\n  - evaluated={:?}, ", st.payloads.as_ref().unwrap().payload_0);
+            if let Some(payloads) = &mut st.payloads {
+                let mut a: Vec<u8> = payloads.payload.clone().into();
+                a.push(2); // TODO: make something random here! (I suggest mutate with bit-level mutations)
+                a.push(2);
+                a.push(2);
+                a[0] = 2;
+                payloads.payload = a.into();
+                log::debug!("Added a payload at path {path:?}.");
+                Ok(())
+            } else {
+                panic!("Should never happen")
+            }
+        } else {
+            Err(Error::Term(
+                "[add_one_payload_randomly] Unable to make_message".to_string(),
+            ))
+        }
+    } else {
+        Err(Error::Term(
+            "[add_one_payload_randomly] Unable to choose a suitable sub-term".to_string(),
+        ))
+    }
 }
